# public/openapi.yaml
openapi: 3.0.0
info:
  title: Online library
  version: 1.0.0
servers:
  - url: https://tim7.ictcortex.me/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              msg:
                type: string
                example: "invalid token"
paths:
  /login:
    post:
      summary: Login with a username and password
      parameters:
        - name: username
          in: query
          description: Username
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: Username
          required: true
          schema:
            type: string
      responses:
        200:
          description: Login success
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "success"
                  plainTextToken:
                    type: string
                    example: "27|Qv1aXddrHhoCRJH9bEkplyQBdtwwIAX9md7178Na"
        401:
          description: Invalid username and/or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "failure"
                  plainTextToken:
                    type: string
                    example: null
  /register:
    post:
      summary: Create a new student account
      parameters:
        - name: username
          in: query
          description: Username
          required: true
          schema:
            type: string
        - name: name
          in: query
          description: Name
          required: true
          schema:
            type: string
        - name: email
          in: query
          description: Email
          required: true
          schema:
            type: string
        - name: jmbg
          in: query
          description: JMBG
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: Username
          required: true
          schema:
            type: string
      responses:
        200:
          description: Registration success
          content:
            application/json:
              schema:
                type: object
                properties:
                  plainTextToken:
                    type: string
                    example: "27|Qv1aXddrHhoCRJH9bEkplyQBdtwwIAX9md7178Na"
  /forgot-username:
    post:
      summary: Request an email containing your username
      parameters:
        - name: email
          in: query
          description: Email
          required: true
          schema:
            type: string
      responses:
        200:
          description: Email was sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "success"
        500:
          description: No user with the given email exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "invalid email"
  /forgot-password:
    post:
      summary: Request an email containing a link to reset your password
      parameters:
        - name: username
          in: query
          description: Username
          required: true
          schema:
            type: string
      responses:
        200:
          description: Email was sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "success"
        500:
          description: No user with the given username exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "invalid email"
  /user:
    get:
      summary: Get current user data
      security:
        - bearerAuth: []
      responses:
        401:
          $ref: '#/components/responses/UnauthorizedError'
        200:
          description: User data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  jmbg:
                    type: string
                    example: "1869507997853"
                  name:
                    type: string
                    example: "Haris Sabotic"
                  email:
                    type: string
                    example: "haris.sinfo16@gmail.com"
                  username:
                    type: string
                    example: "haris"
                  photo:
                    type: string
                    example: "sample1_1660957917.jpeg"
  /edit-user:
    post:
      summary: Edit current user data
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          description: Name
          required: true
          schema:
            type: string
        - name: email
          in: query
          description: Email
          required: true
          schema:
            type: string
        - name: username
          in: query
          description: Username
          required: true
          schema:
            type: string
        - name: oldPass
          in: query
          description: User's current password
          required: true
          schema:
            type: string
        - name: newPass
          in: query
          description: New password for the user
          required: true
          schema:
            type: string
      responses:
        401:
          $ref: '#/components/responses/UnauthorizedError'
        200:
          description: User data changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "success"
        403:
          description: User's current password(oldPass) invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "old password invalid"
  /zahtjevi:
    get:
      summary: Get current user's reservations and rents
      security:
        - bearerAuth: []
      parameters:
        - name: filter
          in: query
          description: Can be "reservations", "rents", "returned" or blank
          schema:
            type: string
      responses:
        401:
          $ref: '#/components/responses/UnauthorizedError'
        200:
          description: Data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    id:
                      type: string
                      example: "1"
                    book:
                      type: object
                      example:
                          title: "Cassidy Rippin"
                          authors: [
                            "Jammie Jones",
                            "Johnnie Cronin",
                            "Johnnie Cronin"
                          ]
                    librarian:
                      type: string
                      example: "Emilija Pribanovic"
                    dateFrom:
                      type: string
                      example: "2022-08-20"
                    dateTo:
                      type: string
                      description: Can be empty in case type is rent, which means the book hasn't been returned
                      example: "2022-08-19"
                    type:
                      type: string
                      description: Can be "reservation rejected", "reservation" or "rent"
                      example: "reservation rejected"
  /izbrisi-transakciju:
    delete:
      summary: Delete one of the current user's reservations or rents
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          description: Reservation or rent's ID
          schema:
            type: string
        - name: type
          in: query
          description: Can be "reservation", "rent", or "reservation rejected"
          schema:
            type: string
      responses:
        401:
          $ref: '#/components/responses/UnauthorizedError'
        200:
          description: Successfully deleted reservation or rent
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "success"
        500:
          description: Tried deleting a rent when the book associated with it hasn't been returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "book not returned"
  /rezervisi:
    post:
      summary: Make a reservation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          description: Book ID
          schema:
            type: string
        - name: dateFrom
          in: query
          description: Reservation's start date
          schema:
            type: string
        - name: dateTo
          in: query
          description: Reservation's end date
          schema:
            type: string
      responses:
        401:
          $ref: '#/components/responses/UnauthorizedError'
        200:
          description: Successfully created reservation
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "success"
  /aktivnosti:
    get:
      summary: Get a sorted list of all activities associated with the current user
      security:
        - bearerAuth: []
      responses:
        401:
          $ref: '#/components/responses/UnauthorizedError'
        200:
          description: Data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    book:
                      type: string
                      example: "Demarco Muller"
                    photo:
                      type: string
                      example: "sample2_1661343898.jpeg"
                    librarian:
                      type: string
                      example: "Emilija Pribanovic"
                    date:
                      type: string
                      example: "2022-08-19 00:00:00"
                    type:
                      type: string
                      description: Can be "reservation" or "rent"
                      example: "reservation"
...
  /books/{book}:
    get:
      summary: Get the details of a specific book
      parameters:
        - in: path
          name: book
          description: Book ID
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 11
                  title:
                    type: string
                    example: "Demarco Muller"
                  authors:
                    type: array
                    items:
                      type: string
                    example: [
                      "Johnnie Cronin",
                      "Erwin Fay"
                    ]
                  summary:
                    type: string
                    example: "Tenetur id quasi blanditiis explicabo."
                  available:
                    type: boolean
                    example: true
                  quantity:
                    type: integer
                    example: 27
                  categories:
                    type: array
                    items:
                      type: string
                    example: []
                  genres:
                    type: array
                    items:
                      type: string
                    example: [
                      "Poetry"
                    ]
                  publisher:
                    type: string
                    example: "BeoBook"
                  publishYear:
                    type: string
                    example: "1647"
                  photo:
                    type: string
                    example: ""
  /book-spec/{book}:
    get:
      summary: Get the specification of a specific book
      parameters:
        - in: path
          name: book
          description: Book ID
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages:
                    type: string
                    example: 33
                  pismo:
                    type: string
                    example: Latin
                  jezik:
                    type: string
                    example: English
                  povez:
                    type: string
                    example: Paperback
                  format:
                    type: string
                    example: A1
                  isbn:
                    type: string
                    example: 126-1-55-439923-0
...